+++ egui/src/context.rs
@@ -48,6 +48,7 @@ struct ContextImpl {
 
     /// While positive, keep requesting repaints. Decrement at the end of each frame.
     repaint_requests: u32,
+    request_repaint_callbacks: Option<Box<dyn Fn() + Send + Sync>>,
 }
 
 impl ContextImpl {
@@ -122,7 +123,7 @@ impl ContextImpl {
 ///
 /// ``` no_run
 /// # fn handle_platform_output(_: egui::PlatformOutput) {}
-/// # fn paint(textures_detla: egui::TexturesDelta, _: Vec<egui::ClippedMesh>) {}
+/// # fn paint(textures_detla: egui::TexturesDelta, _: Vec<egui::ClippedPrimitive>) {}
 /// let mut ctx = egui::Context::default();
 ///
 /// // Game loop:
@@ -137,8 +138,8 @@ impl ContextImpl {
 ///         });
 ///     });
 ///     handle_platform_output(full_output.platform_output);
-///     let clipped_meshes = ctx.tessellate(full_output.shapes); // create triangles to paint
-///     paint(full_output.textures_delta, clipped_meshes);
+///     let clipped_primitives = ctx.tessellate(full_output.shapes); // create triangles to paint
+///     paint(full_output.textures_delta, clipped_primitives);
 /// }
 /// ```
 #[derive(Clone)]
@@ -533,11 +534,28 @@ impl Context {
 
 impl Context {
     /// Call this if there is need to repaint the UI, i.e. if you are showing an animation.
+    ///
     /// If this is called at least once in a frame, then there will be another frame right after this.
     /// Call as many times as you wish, only one repaint will be issued.
+    ///
+    /// If called from outside the UI thread, the UI thread will wake up and run,
+    /// provided the egui integration has set that up via [`Self::set_request_repaint_callback`]
+    /// (this will work on `eframe`).
     pub fn request_repaint(&self) {
         // request two frames of repaint, just to cover some corner cases (frame delays):
-        self.write().repaint_requests = 2;
+        let mut ctx = self.write();
+        ctx.repaint_requests = 2;
+        if let Some(callback) = &ctx.request_repaint_callbacks {
+            (callback)();
+        }
+    }
+
+    /// For integrations: this callback will be called when an egui user calls [`Self::request_repaint`].
+    ///
+    /// This lets you wake up a sleeping UI thread.
+    pub fn set_request_repaint_callback(&self, callback: impl Fn() + Send + Sync + 'static) {
+        let callback = Box::new(callback);
+        self.write().request_repaint_callbacks = Some(callback);
     }
 
     /// Tell `egui` which fonts to use.
@@ -773,7 +791,7 @@ impl Context {
     }
 
     /// Tessellate the given shapes into triangle meshes.
-    pub fn tessellate(&self, shapes: Vec<ClippedShape>) -> Vec<ClippedMesh> {
+    pub fn tessellate(&self, shapes: Vec<ClippedShape>) -> Vec<ClippedPrimitive> {
         // A tempting optimization is to reuse the tessellation from last frame if the
         // shapes are the same, but just comparing the shapes takes about 50% of the time
         // it takes to tessellate them, so it is not a worth optimization.
@@ -782,13 +800,13 @@ impl Context {
         tessellation_options.pixels_per_point = self.pixels_per_point();
         tessellation_options.aa_size = 1.0 / self.pixels_per_point();
         let paint_stats = PaintStats::from_shapes(&shapes);
-        let clipped_meshes = tessellator::tessellate_shapes(
+        let clipped_primitives = tessellator::tessellate_shapes(
             shapes,
             tessellation_options,
             self.fonts().font_image_size(),
         );
-        self.write().paint_stats = paint_stats.with_clipped_meshes(&clipped_meshes);
-        clipped_meshes
+        self.write().paint_stats = paint_stats.with_clipped_primitives(&clipped_primitives);
+        clipped_primitives
     }
 
     // ---------------------------------------------------------------------
@@ -1245,4 +1263,11 @@ impl Context {
         style.ui(ui);
         self.set_style(style);
     }
+}
+
+#[cfg(test)]
+#[test]
+fn context_impl_send_sync() {
+    fn assert_send_sync<T: Send + Sync>() {}
+    assert_send_sync::<Context>();
 }