+++ egui_glow/src/epi_backend.rs
@@ -3,14 +3,6 @@ use egui_winit::winit;
 
 struct RequestRepaintEvent;
 
-struct GlowRepaintSignal(std::sync::Mutex<winit::event_loop::EventLoopProxy<RequestRepaintEvent>>);
-
-impl epi::backend::RepaintSignal for GlowRepaintSignal {
-    fn request_repaint(&self) {
-        self.0.lock().unwrap().send_event(RequestRepaintEvent).ok();
-    }
-}
-
 #[allow(unsafe_code)]
 fn create_display(
     window_builder: winit::window::WindowBuilder,
@@ -47,28 +39,42 @@ pub use epi::NativeOptions;
 
 /// Run an egui app
 #[allow(unsafe_code)]
-pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
-    let persistence = egui_winit::epi::Persistence::from_app_name(app.name());
+pub fn run(app_name: &str, native_options: &epi::NativeOptions, app_creator: epi::AppCreator) -> ! {
+    let persistence = egui_winit::epi::Persistence::from_app_name(app_name);
     let window_settings = persistence.load_window_settings();
     let window_builder =
-        egui_winit::epi::window_builder(native_options, &window_settings).with_title(app.name());
+        egui_winit::epi::window_builder(native_options, &window_settings).with_title(app_name);
     let event_loop = winit::event_loop::EventLoop::with_user_event();
     let (gl_window, gl) = create_display(window_builder, &event_loop);
+    let gl = std::rc::Rc::new(gl);
 
-    let repaint_signal = std::sync::Arc::new(GlowRepaintSignal(std::sync::Mutex::new(
-        event_loop.create_proxy(),
-    )));
-
-    let mut painter = crate::Painter::new(&gl, None, "")
+    let mut painter = crate::Painter::new(gl.clone(), None, "")
         .unwrap_or_else(|error| panic!("some OpenGL error occurred {}\n", error));
+
     let mut integration = egui_winit::epi::EpiIntegration::new(
         "egui_glow",
         painter.max_texture_side(),
         gl_window.window(),
-        repaint_signal,
         persistence,
-        app,
     );
+
+    {
+        let event_loop_proxy = parking_lot::Mutex::new(event_loop.create_proxy());
+        integration.egui_ctx.set_request_repaint_callback(move || {
+            event_loop_proxy.lock().send_event(RequestRepaintEvent).ok();
+        });
+    }
+
+    let mut app = app_creator(&epi::CreationContext {
+        egui_ctx: integration.egui_ctx.clone(),
+        integration_info: integration.frame.info(),
+        storage: integration.persistence.storage(),
+        gl: gl.clone(),
+    });
+
+    if app.warm_up_enabled() {
+        integration.warm_up(app.as_mut(), gl_window.window());
+    }
 
     let mut is_focused = true;
 
@@ -88,15 +94,15 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                 needs_repaint,
                 textures_delta,
                 shapes,
-            } = integration.update(gl_window.window());
+            } = integration.update(app.as_mut(), gl_window.window());
 
             integration.handle_platform_output(gl_window.window(), platform_output);
 
-            let clipped_meshes = integration.egui_ctx.tessellate(shapes);
+            let clipped_primitives = integration.egui_ctx.tessellate(shapes);
 
             // paint:
             {
-                let color = integration.app.clear_color();
+                let color = app.clear_color();
                 unsafe {
                     use glow::HasContext as _;
                     gl.disable(glow::SCISSOR_TEST);
@@ -104,10 +110,9 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                     gl.clear(glow::COLOR_BUFFER_BIT);
                 }
                 painter.paint_and_update_textures(
-                    &gl,
                     gl_window.window().inner_size().into(),
                     integration.egui_ctx.pixels_per_point(),
-                    clipped_meshes,
+                    &clipped_primitives,
                     &textures_delta,
                 );
 
@@ -125,7 +130,7 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                 };
             }
 
-            integration.maybe_autosave(gl_window.window());
+            integration.maybe_autosave(app.as_mut(), gl_window.window());
         };
 
         match event {
@@ -144,7 +149,7 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                     gl_window.resize(physical_size);
                 }
 
-                integration.on_event(&event);
+                integration.on_event(app.as_mut(), &event);
                 if integration.should_quit() {
                     *control_flow = winit::event_loop::ControlFlow::Exit;
                 }
@@ -152,8 +157,11 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                 gl_window.window().request_redraw(); // TODO: ask egui if the events warrants a repaint instead
             }
             winit::event::Event::LoopDestroyed => {
-                integration.on_exit(gl_window.window());
-                painter.destroy(&gl);
+                integration
+                    .persistence
+                    .save(&mut *app, &integration.egui_ctx, gl_window.window());
+                app.on_exit(&gl);
+                painter.destroy();
             }
             winit::event::Event::UserEvent(RequestRepaintEvent) => {
                 gl_window.window().request_redraw();